+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog for Bearded Theme
 
+## 8.1.0 - 2023-04-??
+
+### Enhancements
+
+### Fixes
+
+- Fix border bottom on active tab hover
+- Fix background colors for diffs
+- Fix unwanted border on buttons
+
+### New features
+
 ## 8.0.0 - 2023-03-20
 
 ### Enhancements